gost_crypt: add support for plain CFB and OFB modes
Rationale: The meshing of key and initialization vector makes it
impossible to use stateless protocols.

This change adds the ability to use plain OFB mode without key meshing
and, most importantly, an initialization vector meshing. It is worth
noting that plain CFB mode can be obtained simply by disabling key
meshing, but a plain mode option is also supported, added for
generality.

Note that the CBC mode does not use meshing in any form, so this change
does not affect it.